requirements.txt
# Google Agent Development Kit
google-adk[database]==0.3.0
google-generativeai==0.8.5

# Environment variables management
python-dotenv

# Database dependencies (if not included with google-adk)
sqlalchemy
#sqlite3

# Optional: Development and debugging tools
pytest
black
flake8